LPD-79450 SF

+++ portal-kernel/src/com/liferay/portal/kernel/upgrade/data/cleanup/UserAllTablesOrphanReferencesDataCleanupPreupgradeProcess.java
@@ -100,7 +100,7 @@ public class UserAllTablesOrphanReferencesDataCleanupPreupgradeProcess
 
 			while (resultSet.next()) {
 				long companyId = resultSet.getLong("companyId");
-				long count = resultSet.getLong("count");
+				long count = resultSet.getInt("count");
 				long userId = resultSet.getLong(sourceColumnName);
 
 				if (_deleteTableNames.contains(sourceTableName) ||
